Added parameter serializers for more modular code.
